refactored test files
