+++ query-service-impl/src/test/java/org/hypertrace/core/query/service/pinot/ExecutionContextTest.java
@@ -1,7 +1,7 @@
 package org.hypertrace.core.query.service.pinot;
 
-import static org.hypertrace.core.query.service.QueryRequestUtil.createFilter;
-import static org.hypertrace.core.query.service.QueryRequestUtil.createTimeColumnGroupByFunction;
+import static org.hypertrace.core.query.service.QueryRequestBuilderUtils.createFilter;
+import static org.hypertrace.core.query.service.QueryRequestBuilderUtils.createTimeColumnGroupByFunction;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
